Cleanup Universal System Warnings (#1596)
* The motion planning and path extensions had misleading indentation on a for loop that reallocates their data structure. In both cases, the outer array is being copied to a new outer array, but with the same internal pointers. So I figured out that the delete was intended to go on a new line, outside the for loop, because, again, we're only deleting the outer array not the inner rows.
* Fixed warning in `script_exists` by casting the script id to size_t so it can be safely compared with `enigma::script_idmax`.
* Fixed warnings in `room_next`, `room_previous`, and `room_exists` by casting the room id to size_t so it can be safely compared with `enigma::room_idmax`.
* Changed the control variable in `array_equals` from size_t to int so that it can be safely compared to the `array_len()` member of Josh's VLAs.
* Removed two of Josh's unused variables in the zlib sources that I could not find any other reference to.

+++ ENIGMAsystem/SHELL/Universal_System/resource_data.cpp
@@ -79,8 +79,7 @@ variant script_execute(int scr, variant arg0, variant arg1, variant arg2, varian
 
 bool script_exists(int script)
 {
-    return (script >= 0 && script < enigma::script_idmax && bool(enigma::callable_scripts[script].base));
+    return (script >= 0 && size_t(script) < enigma::script_idmax && bool(enigma::callable_scripts[script].base));
 }
 
 }
-
