+++ ENIGMAsystem/SHELL/Universal_System/roomsystem.cpp
@@ -81,20 +81,20 @@ namespace enigma
   {
     using namespace enigma_user;
     object_basic *instanceexists = fetch_instance_by_int(vob);
-  
+
     if (instanceexists)
     {
       object_planar* vobr = (object_planar*)instanceexists;
-  
+
       double vobx = vobr->x, voby = vobr->y;
-  
+
       //int bbl=*vobr.x+*vobr.bbox_left,bbr=*vobr.x+*vobr.bbox_right,bbt=*vobr.y+*vobr.bbox_top,bbb=*vobr.y+*vobr.bbox_bottom;
       //if (bbl<view_xview[vc]+view_hbor[vc]) view_xview[vc]=bbl-view_hbor[vc];
-  
+
       double vbc_h, vbc_v;
       (view_hborder[vc] > view_wview[vc]/2) ? vbc_h = view_wview[vc]/2 : vbc_h = view_hborder[vc];
       (view_vborder[vc] > view_hview[vc]/2) ? vbc_v = view_hview[vc]/2 : vbc_v = view_vborder[vc];
-  
+
       if (view_hspeed[vc] == -1)
       {
         if (vobx < view_xview[vc] + vbc_h)
@@ -117,7 +117,7 @@ namespace enigma
             view_xview[vc] = vobx + vbc_h - view_wview[vc];
         }
       }
-  
+
       if (view_vspeed[vc] == -1)
       {
         if (voby < view_yview[vc] + vbc_v)
@@ -140,19 +140,19 @@ namespace enigma
             view_yview[vc] = voby + vbc_v - view_hview[vc];
         }
       }
-  
+
       if (view_xview[vc] < 0)
         view_xview[vc] = 0;
       else if (view_xview[vc] > room_width - view_wview[vc])
         view_xview[vc] = room_width - view_wview[vc];
-  
+
       if (view_yview[vc] < 0)
         view_yview[vc] = 0;
       else if (view_yview[vc] > room_height - view_hview[vc])
         view_yview[vc] = room_height - view_hview[vc];
     }
   }
-  
+
 
   void roomstruct::end() {
     // Fire the Room End event.
@@ -427,7 +427,7 @@ int room_goto_previous()
 
 int room_next(int num)
 {
-  if (num < 0 or num >= enigma::room_idmax)
+  if (num < 0 or size_t(num) >= enigma::room_idmax)
     return -1;
   enigma::roomstruct *rit = enigma::roomdata[num];
   if (!rit or rit->order+1 >= enigma::room_loadtimecount)
@@ -436,7 +436,7 @@ int room_next(int num)
 }
 int room_previous(int num)
 {
-  if (num < 0 or num >= enigma::room_idmax)
+  if (num < 0 or size_t(num) >= enigma::room_idmax)
     return -1;
   enigma::roomstruct *rit = enigma::roomdata[num];
   if (!rit or rit->order-1 < 0)
@@ -446,7 +446,7 @@ int room_previous(int num)
 
 bool room_exists(int roomid)
 {
-  return roomid >= 0 and roomid < enigma::room_idmax and enigma::roomdata[roomid];
+  return roomid >= 0 and size_t(roomid) < enigma::room_idmax and enigma::roomdata[roomid];
 }
 
 int room_set_width(int indx, int wid)
