+++ ENIGMAsystem/SHELL/Universal_System/var_array.cpp
@@ -34,7 +34,7 @@ var array_create(size_t size, variant value) { return var(value, size); }
 var array_create_2d(size_t length, size_t height, variant value) { return var(value, length, height); }
 bool array_equals(const var& arr1, const var& arr2) {
   if (arr1.array_len() != arr2.array_len()) return false;
-  for (size_t i = 0; i < arr1.array_len(); i++) {
+  for (int i = 0; i < arr1.array_len(); i++) {
     if (arr1[i] != arr2[i]) return false;
   }
   return true;
